split tracker
